add application trait
