+++ src/lib.rs
@@ -51,21 +51,11 @@
     clippy::default_trait_access
 )]
 
-use std::{
-    any, fmt,
-    marker::PhantomData,
-    mem,
-    path::Path,
-    rc::Rc,
-    sync::atomic::{AtomicBool, Ordering},
-};
+use std::{any, fmt, marker::PhantomData, path::Path, rc::Rc};
 
 use static_assertions::assert_not_impl_any;
 
-use glutin::{
-    event_loop::EventLoop,
-    window::{Window, WindowBuilder},
-};
+use glutin::event_loop::EventLoop;
 
 use image::RgbaImage;
 
@@ -85,8 +75,10 @@ macro_rules! bug {
 }
 
 mod backend;
-pub mod color;
+mod context;
 mod error;
+
+pub mod color;
 pub mod target;
 
 pub use error::*;
@@ -199,7 +191,16 @@ impl<T: DrawTarget> DrawTarget for &mut T {
     }
 }
 
-static INITIALIZED: AtomicBool = AtomicBool::new(false);
+/// This is the core trait of this library.
+pub trait Application {
+    /// This function is called once per `frame`,
+    /// returning `true` stops the application.
+    fn frame(&mut self, ctx: &mut Context, surface: &mut WindowSurface) -> bool;
+
+    /// This function is called after `frame` has returned `true`
+    /// and can be used to cleanup ressources.
+    fn shutdown(self);
+}
 
 /// A struct storing the global state which is used
 /// for all operations which require access to the GPU.
@@ -244,210 +245,10 @@ static INITIALIZED: AtomicBool = AtomicBool::new(false);
 #[derive(Debug)]
 pub struct Context {
     backend: Backend,
+    event_loop: Option<EventLoop<()>>,
 }
 
 assert_not_impl_any!(Context: Send, Sync, Clone);
-
-impl Context {
-    /// Creates a new `Context`. It is not possible to have more
-    /// than one `Context` at a time.
-    ///
-    /// To create a new `Context` after a previous context was used,
-    /// The previous context has to be dropped using the method
-    /// `Context::unlock_unchecked()`. This is a workaround and
-    /// will probably be fixed in a future release.
-    pub fn new(window: WindowBuilder) -> Result<Self, NewContextError> {
-        if INITIALIZED.compare_and_swap(false, true, Ordering::AcqRel) {
-            panic!("Tried to initialize a second Context");
-        }
-
-        let backend = Backend::initialize(window)?;
-        Ok(Self { backend })
-    }
-
-    /// Returns the dimensions of the used window.
-    pub fn window_dimensions(&self) -> (u32, u32) {
-        self.backend.window_dimensions()
-    }
-
-    /// Returns the width of the used window.
-    pub fn window_width(&self) -> u32 {
-        self.window_dimensions().0
-    }
-
-    /// Returns the height of the used window.
-    pub fn window_height(&self) -> u32 {
-        self.window_dimensions().1
-    }
-
-    /// Sets the dimensions of the used window.
-    pub fn resize_window(&mut self, width: u32, height: u32) {
-        self.backend.resize_window(width, height)
-    }
-
-    /// Returns a handle to the window surface which can be used
-    /// in [`Context::draw`] to draw to the window.
-    ///
-    /// [`Context::draw`]: struct.Context.html#method.draw
-    pub fn window_surface(&self) -> WindowSurface {
-        WindowSurface {
-            _marker: PhantomData,
-        }
-    }
-
-    /// Returns the size of the biggest supported texture.
-    ///
-    /// Trying to create a texture with a size
-    /// greater than `maximum_texture_size` results in an
-    /// `InvalidTextureSize` error.
-    ///
-    /// ```rust, no_run
-    /// use crow::{Context, glutin::WindowBuilder};
-    ///
-    /// let mut ctx = Context::new(WindowBuilder::new()).unwrap();
-    /// println!("maximum supported texture size: {:?}", ctx.maximum_texture_size());
-    /// ```
-    pub fn maximum_texture_size(&self) -> (u32, u32) {
-        self.backend.constants().max_texture_size
-    }
-
-    /// Draws the `source` onto `target`.
-    ///
-    /// To draw to the window, use [`Context::window_surface`] as a target.
-    ///
-    /// [`Context::window_surface`]: struct.Context.html#method.window_surface
-    pub fn draw<T>(
-        &mut self,
-        target: &mut T,
-        source: &Texture,
-        position: (i32, i32),
-        config: &DrawConfig,
-    ) where
-        T: DrawTarget,
-    {
-        target.receive_draw(self, source, position, config)
-    }
-
-    /// Draws the a line going from `from` to `to` onto `target` with the given `color`.
-    ///
-    /// To draw this line to the window, use [`Context::window_surface`] as a target.
-    ///
-    /// [`Context::window_surface`]: struct.Context.html#method.window_surface
-    pub fn debug_line<T>(
-        &mut self,
-        target: &mut T,
-        from: (i32, i32),
-        to: (i32, i32),
-        color: (f32, f32, f32, f32),
-    ) where
-        T: DrawTarget,
-    {
-        target.receive_line(self, from, to, color)
-    }
-
-    /// Draws the bounding box of an axis-aligned rectangle specified by
-    /// its `lower_left` and `upper_right` corner.
-    ///
-    /// In case `lower_left` is to the right or above `upper_right`, the two points will be flipped.
-    ///
-    /// To draw this rectangle to the window, use [`Context::window_surface`] as a target.
-    ///
-    /// [`Context::window_surface`]: struct.Context.html#method.window_surface
-    pub fn debug_rectangle<T>(
-        &mut self,
-        target: &mut T,
-        lower_left: (i32, i32),
-        upper_right: (i32, i32),
-        color: (f32, f32, f32, f32),
-    ) where
-        T: DrawTarget,
-    {
-        target.receive_rectangle(self, lower_left, upper_right, color)
-    }
-
-    /// Clears the color of the given [`DrawTarget`], setting each pixel to `color`
-    ///
-    /// [`DrawTarget`]: trait.DrawTarget.html
-    pub fn clear_color<T>(&mut self, target: &mut T, color: (f32, f32, f32, f32))
-    where
-        T: DrawTarget,
-    {
-        target.receive_clear_color(self, color)
-    }
-
-    /// Resets the depth buffer of the given [`DrawTarget`] to `1.0`.
-    ///
-    /// [`DrawTarget`]: trait.DrawTarget.html
-    pub fn clear_depth<T>(&mut self, target: &mut T)
-    where
-        T: DrawTarget,
-    {
-        target.receive_clear_depth(self)
-    }
-
-    /// Stores the current state of the window in an image.
-    /// This function is fairly slow and should not be used carelessly.
-    ///
-    /// It is currently not possible to screenshot a part of the screen.
-    pub fn take_screenshot(&mut self) -> RgbaImage {
-        let (width, height) = self.window_dimensions();
-
-        let data = self.backend.take_screenshot((width, height));
-
-        let reversed_data = data
-            .chunks(width as usize * 4)
-            .rev()
-            .flat_map(|row| row.iter())
-            .copied()
-            .collect();
-
-        RgbaImage::from_vec(width, height, reversed_data).unwrap()
-    }
-
-    /// Returns the inner window.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use crow::{Context, glutin::WindowBuilder};
-    ///
-    /// let context = Context::new(WindowBuilder::new().with_title("Starting"))
-    ///     .expect("unable to create a context");
-    ///
-    /// context.window().set_title("Running");
-    /// ```
-    pub fn window(&self) -> &Window {
-        self.backend.window()
-    }
-
-    /// Hijacks the calling thread and runs the given `frame` in a loop.
-    ///
-    /// Since the closure is `'static`, it must be a `move` closure if it needs to
-    /// access any data from the calling context.
-    pub fn run<F>(self, frame: F) -> ! {
-        self.backend
-            .run(move |backend: &mut Backend, (event, window_target, control_flow)| {
-
-            })
-    }
-
-    /// Drops this context while allowing the initialization of a new one afterwards.
-    ///
-    /// # Safety
-    ///
-    /// This method may lead to undefined behavior if a struct, for example a `Texture`, which was created using
-    /// the current context, is used with the new context.
-    pub unsafe fn unlock_unchecked(self) {
-        mem::drop(self);
-
-        let gl_error = gl::GetError();
-        if gl_error != gl::NO_ERROR {
-            bug!("unexpected error: {}", gl_error);
-        }
-
-        INITIALIZED.store(false, Ordering::Release);
-    }
-}
 
 /// A handle which can be used to draw to the window.
 ///
